Show the current or the old value according the change's state
The previous implementation had an issue which was that when a revision was
created, the partner's value was copied to the revision change and so if the
partner was edited, the revision change could show an outdated value.

What we want to ensure is to show the actual value when the change is pending
and show the old value when the change is done/canceled.