crm_deduplicate_filter
===========================================
Exclude some records from the deduplication
===========================================

This module extends the possibilities of the contact deduplication allowing
to filter the applicable set according several criteria.

For now, only the filter for restricting the deduplication to only companies or
only contacts is implemented.

Usage
=====

To use this module, you need to:

   "Parent company set (Contacts)" in the section 'Exclude contacts having'.
   kind of records.

OCA Transbot updated translations from Transifex