[ADD] partner_default_sale_discount
Default sales discount per partner
==================================

This module allows to define at partner level a default discount to be applied
on sales orders as the default one. As the field is a company dependent one,
you can configure a different value for the partner for each of your companies.

Configuration
=============

Enabling the use of discounts in Odoo:

* Go to *Sales > Configuration > Settings*
* On the group "Quotations & Sales", select the option "Allow discounts on
  sales order lines" for the "Discount" section.

Setting a default sales discount:

* Then, go to *Sales > Customers*.
* Create or modify a customer.
* Go to "Sales & Purchases" page.
* In the "Sale" section, there's a field called "Default sales discount (%)"
  where you can fill the default discount percentage.
* This value can only be filled at company partners, not for contacts.

Usage
=====

To use this module, you need to:

* Go to Sales > Quotations.
* Create a new quotation.
* Select a partner with a default discount applied.
* Create an order line, and you will see this discount as default in the line.

Known issues / Roadmap
======================

* This module doesn't work with *product_visible_discount*, as it overwrites
  the discount field with its own value.