[IMP] partner_noncommercial. Reorganize menu's.
There were menu-options for partners and relations interspersed, and no clear separation between
normal menu's and configuration. This commit tries to rectify this.

I abandoned the idea for 10.0 of bringing everything under the contact menu. In 11.0 relations
are under the contact menu and the dependency on sales_team can be easily removed.

But for 10.0 I would not confuse existing users and installations by changing the dependency graph of
modules and putting stuff under a new main menu (contacts instead of partners).

New options have been translated into Dutch. The rest can be done through weblate.