[FIX] city_id no longer shown.
We want to use a domain that requires city_id to be on the view
but we can't add it directly there, otherwise _fields_view_get_address
in base_address_city won't do its magic, as it immediately returns
if city_id is already in there.

On the other hand, if city_id is not in the
views, odoo won't let us use it in zip_id's domain.

Fixes a bug introduced here: 5fd93fb8d138cb9a650410d49745b0777be43be0