[MIG] partner_deduplicate_acl: Migration to v14.0
[MIG] partner_deduplicate_acl: Migration to 15.0

[FIX] partner_deduplicate_acl: Put menu on lower sequence

So that when clicking on Contacts main menu, it doesn't open the
wizard by default.

[FIX] partner_deduplicate_acl: Change menu group

The menu for deduplicating was only accessible for the users having the
group "Automatically", so the group "Manual" users weren't able to access
it.

This commits changes to this second group for both being able to access