[REF] partner_second_lastname: clean firstname and lastname2 when partner is a company
When you create a partner type individual and update to company,
the field `lastname` is filled-out with the full name but the fields
`firstname` and lastname2 are not cleaned,
it's causing when you update the partner name with a different order,
the full name is computed based on the
fields `firstname`, `lastname` and lastname2 and it's inconsistent.

This commit is cleaning the fields `firstname` and `lastname2` from the
write method.