[ADD] module partner_child_ids_confirm_remove
This module add a simple confirm dialog before the child id partner deletion: without an explicit warning on delete button it could be possible to click it without even notice that partner is going to be deleted. Saving the record would then lead to unintentional loss of data. About the implementation, it would be much better to properly patch the on_remove() function instead of mocking the _remove property as I did, but I'm not too into JS and I did not find a good way to do it. The fact that I am not patching on_remove() function means that it could break inheritance by 3rd party modules, but the _remove prop is not inherited in the source code, so the native behaviour will not be compromised. Possible improvements by good JS developers.