[IMP] partner_pricelist_search: support searching subfields
Previously we couldn't make an advanced search to filter pricelist
subfields.

Currently, only possitive operators are working for advanced searches as
negative ones seem to be encapsulating the queries and we can't use them
in the point we're overriding. Further research would need to be done in
order to cover it.

TT47648