[FIX] partner_manual_rank: Compatibility with Odoo default
The introduction of default values for the fields 'is_customer' and 'is_supplier' breaks the compatibility when the create method on partner is called with once of the fields 'supplier_rank', 'customer_rank'. To restore this compatibility, we ovveride the create method to initialize the 'is_customer' and 'is_supplier' fields into the given dict if 'supplier_rank' or 'customer_rank' are passed to the method.