[REM] partner_company_group: deprecated module
This module is a left-over from a refactor done in 15.0:
- https://github.com/OCA/partner-contact/pull/1347
-> https://github.com/OCA/partner-contact/commit/7f3f531e

This refactor split the module in 4:
- base_partner_company_group
- account_partner_company_group
- crm_partner_company_group
- sale_partner_company_group

The purpose is to clearly allow the installation of the "Company Group" feature
without pulling all the complete applications like Sales or CRM.

Then, this module was left as an easy migration path for people that had the
module installed in 15.0 previous to this refactor, with XML functions to update
the XML ids of the previous views to the ones in the new modules.

As such, it should not have been migrated to the newer versions.