[FIX] base_partner_sequence: assign a new ref when copying users
Port forward of https://github.com/OCA/partner-contact/pull/2057

When copying a user that has a partner that is a company and that has a ref,
the resulting new user was assigned a new partner but with the same ref.

This is an issue for instance for new users signing up, if the partner of the
template user is a company and is assigned a reference.

Steps to reproduce the behavior:
1. Copy a user that has a partner that is a company and has a ref.

**Expected behavior**
The new user has a partner with a newly assigned reference.